requirements.txt
paramiko==2.7.1
boto3==1.12.14
PyYAML==5.3
azure-common==1.1.26
azure-core==1.9.0
azure-identity==1.5.0
azure-mgmt-compute==18.0.0
azure-mgmt-core==1.2.2
azure-mgmt-network==16.0.0
